rdma-core-18

Updates from version 17 (git log):

 * buildlib: Properly update github release with pandoc
 * debian: Bump Standards-Version to 4.1.4
 * debian: Whitelist coherent DMA supporting architectures
 * debian: Whitelist architectures that support valgrind
 * verbs: Fix man error in attr argument of ibv_alloc_parent_domain
 * buildlib: Fix broken release script
 * ibacm: Adding new configuration option 'acme_plus_kernel_only'
 * ibacm: Adding new configuration option 'server_mode'
 * Check for broken strict aliasing
 * debian: Do not require valgrind on riscv64
 * libibumad/man/umad_open_port.3: Update return values
 * libibumad/man/umad_init.3.md: Trivial punctuation changes
 * umad: Do not check for umad sysfs files in umad_init
 * umad: Switch umad_init.3 to markdown
 * Add a script to update the kernel headers
 * mlx5: Use kernel headers for the mmap command number
 * verbs: Add missing MLX5_1.5 to ibverbs-providers.symbols
 * mlx5: Add mlx5dv support for device memory
 * mlx5: Adding support for device minimum inline header requirement
 * verbs: Add device memory support in rc_pingpong example
 * verbs: Fix wrong clean up flow in ibv_rc_pingpong
 * mlx5: Device memory support
 * Always enable large file support when building
 * verbs: Add device memory (DM) support in libibverbs
 * verbs: Add device memory capabilities to query_device_cap_ex
 * verbs: Add missing documentation on IBV_ACCESS_ZERO_BASED MRs
 * Update kernel headers
 * libibverb/examples: Protect against memory corruption
 * verbs: Have DECLARE_DRV_CMD check that ex commands are 8 byte aligned
 * verbs: Fix mis-spelling of ibv_create_qp_ex_to_reg
 * Update IBVERBS_PABI_VERSION to 18
 * cbuild: Make cbuild work with git worktrees
 * ibacm: empty port provider causes segfault
 * verbs: Add ibv_flow_spec_action_handle support
 * mlx5: Query the FLOW_ACTION capabilities through DV
 * verbs: Add flow steering specification for ESP
 * verbs: Add flow steering egress support
 * verbs: Add ability to hash according to ipsec_spi
 * mlx5: Add mlx5dv_create_flow_action_esp() DV API
 * mlx5: Add support to creating/modifying and destroying ESP flow_action
 * verbs: Add a man page for flow_action_esp
 * verbs: Add flow_action create, modify and destroy verbs
 * verbs: Add flow_action uapi definitions
 * verbs: Add enum attribute type to the ioctl infrastructure
 * rxe: Update abi_version
 * cxgb3: Move cxgb3 to exclusively use the kernel uapi headers
 * cxgb3: Use kernel uapi headers
 * ocrdma: Move ocrdma to exclusively use the kernel uapi headers
 * ocrdma: Use kernel uapi headers
 * hns: Move hns to exclusively use the kernel uapi headers
 * i40iw: Move i40iw to exclusively use the kernel uapi headers
 * i40iw: Use kernel uapi headers
 * bnxt: Move bnxt to exclusively use the kernel uapi headers
 * bnxt_re: Use kernel uapi headers
 * vmw_pvrdma: Use new kABI macros
 * mthca: Move mthca to exclusively use the kernel uapi headers
 * mthca: Use kernel uapi headers
 * cxgb4: Move cxgb4 to exclusively use the kernel uapi headers
 * cxgb4: Use kernel uapi headers
 * nes: Move nes to exclusively use the kernel uapi headers
 * nes: Use kernel uapi headers
 * mlx5: Move mlx5 to exclusively use the kernel uapi headers
 * mlx5: Do not use internal ABI structures in place of kernel ABI
 * mlx5: Use kernel uapi header structs
 * mlx5: Use kernel uapi headers
 * mlx4: Move mlx4 to exclusively use the kernel uapi headers
 * mlx4: Use kernel uapi headers
 * rxe: Move rxe to exclusively use the kernel uapi headers
 * qedr: Move qedr to exclusively use the kernel uapi headers
 * mlx4: Enable WQ creation with IBV_WQ_FLAGS_SCATTER_FCS
 * qedr: Update MAINTAINERS
 * suse: librdmacm1 provides/obsoletes librdmacm
 * suse: Build with curl-mini by default
 * suse: fix double %%
 * libhns: Support cq record doorbell
 * libhns: Support rq record doorbell
 * Add driver-id to ioctl() infrastructure
 * verbs: Change ioctl uapi header files
 * verbs: Switch simpler man pages over to markdown format
 * Support writing man pages in MarkDown
 * mlx5: Support ibv_modify_qp_rate_limit
 * Fix typo in rdma_create_symlink
 * verbs: Extend verbs packet pacing QP rate limiting with burst info
 * verbs: Use kabi for create_qp
 * verbs: Replace more of kern-abi.h with kernel headers
 * verbs: Provide infrastructure to remove kernel headers copies
 * verbs: Remove bogus padding from ibv_dealloc_mw
 * Allow static_assert to be used
 * rxe: Get rid of rxe_get_context_resp
 * verbs: Use kabi for IB_UVERBS_CQ_FLAGS_TIMESTAMP_COMPLETION
 * librdmacm: Add command line option to specify server
 * libqedr: fix wc wr_id initialization on failure cases
 * mlx5: Fix referencing uninitialized var in create_qp error flow
 * libhns: Remove the unnecessary type conversion
 * libhns: Update the algorithm for computing queue buffer
 * libhns: Fix QP state judgement before receiving work requests
 * libhns: Fix QP state judgement before sending work requests
 * ccan: Add bitmap support for rdma-core
 * Ensure all public headers compile in C++ mode
 * providers/i40iw: Skip CQEs for destroyed QP
 * providers/i40iw: Remove setting of rem_addr.len
 * cxgb4: add new device pci ids.
 * Update library version to be 18.0
 * mlx4: Add TSO support
 * mlx4: Align query device response
 * ibacm: Support FDR and EDR rates
 * ibacm: Log using RCF3339 timestamps
 * debian: Let libibverbs1 17 break ibverbs-providers 16
 * Fix typo "interchangable" -> "interchangeable"
 * debian: Override package-does-not-install-examples
 * debian: Update copyright for kernel-headers/*
 * Have travis test all three IOCTL_MODE's
 * verbs: Remove _RESERVED enums related to CREATE_CQ
 * verbs: Migrate ibv_cmd_destroy_cq and ibv_cmd_create_cq(_ex) to new kabi
 * verbs: Add basic infrastructure for mixed write and ioctl cmds
 * verbs: Add basic infrastructure support for the kabi ioctl
 * verbs: Fix size of verbs_ex_private allocation

+++ providers/hns/hns_roce_u_hw_v1.c
@@ -411,7 +411,7 @@ static int hns_roce_u_v1_poll_cq(struct ibv_cq *ibvcq, int ne,
 
 	if (npolled) {
 		if (dev->hw_version == HNS_ROCE_HW_VER1) {
-			*cq->set_ci_db = (unsigned short)(cq->cons_index &
+			*cq->set_ci_db = (cq->cons_index &
 					 ((cq->cq_depth << 1) - 1));
 			mmio_ordered_writes_hack();
 		}
